Remove useless call to LoadLocation
This call appears to set the time zone to UTC. However, it actually does nothing.

There _is_ a way to globally set the time zone in GoLang - see here: https://stackoverflow.com/questions/54363451/setting-timezone-globally-in-golang

I do not believe we should do this, and instead rely on system time as
specified by the `TZ` environment variable.

Since all time internally is represented as a unix epoch, time zones
only make a difference as to how it is displayed, and it makes sense to
leave this as the user's local time zone.